allocrunner: Basic test alloc runner
